pyproject.toml
[project]
name = "kuakua-agent-backend"
version = "0.1.0"
description = "Backend API scaffold for Kuakua Agent"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110,<1.0",
  "starlette<1.0",
  "uvicorn[standard]>=0.27,<1.0",
  "pydantic-settings>=2.2,<3.0",
  "httpx>=0.27,<1.0",
  "numpy>=1.26,<3.0",
  "kokoro",
  "misaki[zh]",
  "aiosqlite>=0.20,<1.0",
  "pytest-asyncio>=0.23,<1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["kuakua_agent*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-p no:cacheprovider"
